requirements.txt
gym==0.23.0
matplotlib==3.8.2
numpy==2.2.3
Pillow==10.1.0
Pillow==11.1.0
setuptools==69.0.2
torch==2.6.0
